主要修改 所有的默认图片 和几个bug修复

+++ app/models/education.rb
@@ -33,7 +33,7 @@ class Education < ActiveRecord::Base
   end
 
   def after_destroy
-    self.school.destroy
+    self.school && self.school.destroy
   end
 
 end
